Use lowercase "name" as not to be confusing about what this does
The `use` method assumes you pass in either a class or a "name" as a
symbol along with an evaluated block. The examples I've seen
(https://github.com/slim-template/slim/blob/5025afda23ff617071079648237c8209a33c8f6f/lib/slim/engine.rb#L37)
tend to use an uppercase name which is confusing along side things like
`generator :ArrayBuffer` which serve as shortcuts to the classes. To be
clear that this implementation is _not_ a shortcut to the
Temple::Generator class, I went with a lowercase name.