Pass data to ratchet in explicit `data` variable
It worked fine enough with Tilt the template to assume that the view
context was a hash. However, this is not true in Rails. To make the API
consistent, ratchet assumes data is accessed in a local `data` variable.
For the sake of tests, we have to pass along a context that responds to
`data`.