Update attributes implementation to pass existing to block as string
This prevents us from being able to merge existing attribute data with
given, but I think that's a fair trade off to keep a clear separation of
dynamic and static attribute data. We'll see how it holds up with time.

The main concern I have is for attributes like `class` is that a pattern
might arise where some stylistic classes are added on the front-end and
other behavioral classes are bound from the back. We'll see what
happens...